+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `Ezekiel` will be documented in this file.
 
+## 0.1.1 - 2026-06-21
+
+### Added
+- `whereLike`, `orWhereLike` in `Model`
+
+### Changed
+- Conform `orWhere` parameters with `where`
+
 ## 0.1.0 - 2026-05-24
 
 ### Added